+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to **NCronJob** will be documented in this file. The project
 
 - Teach `JobOptionBuilder` to only accept parameters. By [@nulltoken](https://github.com/nulltoken) in [#241](https://github.com/NCronJob-Dev/NCronJob/pull/241).
 - Teach `JobOptionBuilder` to allow configuration of a startup job. By [@nulltoken](https://github.com/nulltoken) in [#289](https://github.com/NCronJob-Dev/NCronJob/pull/289).
+- Make `DependencyBuilder` expose a non generic `RunJob` overload. By [@nulltoken](https://github.com/nulltoken) in [#293](https://github.com/NCronJob-Dev/NCronJob/pull/293).
 
 ### Changed